# document-mgmt-system/darshan-mgmt-service/swagger-darshan-mgmt-service.yaml
{
   "swagger":"2.0",
   "info":{
      "version":"18.10.0",
      "title":"Darshan Management Service API"
   },
   "basePath":"/api/darshan-service",
   "tags":[
      {
         "name":"DarshanMgmt"
      }
   ],
   "schemes":[
      "https"
   ],
   "paths":{
      "/darshan":{
         "post":{
            "tags":[
               "DarshanMgmt"
            ],
            "summary":"Darshan Saving Service",
            "description":"Darshan Saving Service",
            "operationId":"saveDarshan",
            "consumes":[
               "application/json",
               "application/xml"
            ],
            "produces":[
               "application/json",
               "application/xml"
            ],
            "parameters":[
               {
                  "name":"xChannel",
                  "in":"header",
                  "required":false,
                  "type":"string",
                  "minLength":3,
                  "maxLength":36
               },
               {
                  "in":"body",
                  "name":"body",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Darshan"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/Darshan"
                  }
               },
               "400":{
                  "description":"Bad Request"
               },
               "500":{
                  "description":"Internal Server Error"
                }
            }
         }
      },
    "/darshan/createWithArray":{
    "post":{
      "tags":[
        "DarshanMgmt"
      ],
      "summary":"Darshan saving with input arrays Service",
            "description":"Darshan Finding Service",
            "operationId":"saveDarshansWithArrayInput",
       "consumes":[
               "application/json"
            ],
            "produces":[
               "application/json"
            ],
      "parameters":[
               {
                  "name":"xChannel",
                  "in":"header",
                  "required":false,
                  "type":"string",
                  "minLength":3,
                  "maxLength":36
               },
               {
                  "in":"body",
                  "name":"body",
                  "required":true,
                  "schema":{
           "type" :"array",
           "items":{
            "$ref":"#/definitions/Darshan"
           }
                  }
               }
            ],
      "responses":{
               "200":{
                  "description":"Successfully found"
                 
               },
               "500":{
                  "description":"Internal Server Error"
               }
            }
    }
    },
    "/darshan/{id}":{
         "get":{
            "tags":[
               "DarshanMgmt"
            ],
            "summary":"Find Darshan Details based on the darshan id",
            "description":"Find Darshan Details based on the darshan id",
            "operationId":"findDarshanDetailsById",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "application/json"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "required":true,
                  "description":"Darshan Id",
                  "type":"integer"
               }
            ],
            "responses":{
               "200":{
                  "description":"Successfully found",
                  "schema":{
                     "$ref":"#/definitions/Darshan"
                  }
               },
               "400":{
                  "description":"Invalid Darshan name and password supplied"
               },
               "404":{
                  "description":"Darshan not found or inactive"
               },
               "500":{
                  "description":"Internal Server Error"
               }
            }
         },
      "put":{
            "tags":[
               "DarshanMgmt"
            ],
            "summary":"Update Darshan Details based on the darshan id",
            "description":"Update Darshan Details based on the darshan id",
            "operationId":"updateDarshanById",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "application/json"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "required":true,
                  "description":"Darshan Id",
                  "type":"integer"
               },
         {
                  "in":"body",
                  "name":"body",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Darshan"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Successfully found",
                  "schema":{
                     "$ref":"#/definitions/Darshan"
                  }
               },
               "400":{
                  "description":"Invalid Darshan name supplied"
               },
               "404":{
                  "description":"Darshan not found or inactive"
               },
               "500":{
                  "description":"Internal Server Error"
               }
            }
         },
     "delete":{
            "tags":[
               "DarshanMgmt"
            ],
            "summary":"Darshan Deletion Service based on the darshan id",
            "description":"Darshan Deletion Service based on the darshan id",
            "operationId":"deleteDarshanById",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "application/json"
            ],
            "parameters":[
                {
                   "name":"id",
                  "in":"path",
                  "required":true,
                  "description":"Darshan Id",
                  "type":"integer"
               }
            ],
            "responses":{
               "200":{
                  "description":"Darshan deleted succussfully"
               },
               "400":{
                  "description":"Invalid User id supplied"
               },
               "404":{
                  "description":"Darshan Id not found"
               },
               "500":{
                  "description":"Internal Server Error"
               }
            }
         }
      }
   },
   "definitions":{
      "Darshan":{
         "required":[
            "visitorName"
         ],
         "properties":{
            "id":{
               "type":"integer",
               "format":"int64",
               "description":"Darshan id"
            },
			"visitingDate":{
               "type":"string",
               "format":"date",
               "description":"Visiting date"
            },
			"visitorType":{
               "type":"string",
               "enum": ["MEMBER","SEVADAR,","DEVOTEE"],
               "description":"Visitor Type"
            },
            "visitorId":{
               "type":"string",
               "minLength":1,
               "maxLength":50,
               "description":"Visitor Id"
            },
            "visitorName":{
               "type":"string",
               "minLength":1,
               "maxLength":255,
               "description":"Visitor name"
            },
			"visitorCity":{
               "type":"string",
               "minLength":1,
               "maxLength":255,
               "description":"Visitor City"
            },
			"visitorState":{
               "type":"string",
               "minLength":1,
               "maxLength":250,
               "description":"Visitor State"
            },
			"visitorMobileNo":{
               "type":"string",
               "minLength":1,
               "maxLength":15,
               "description":"Visitor Mobile Number"
            },
			"reference":{
               "type":"string",
               "minLength":1,
               "maxLength":50,
               "description":"Visitor Reference"
            },
			"totalMen":{
               "type":"integer",
               "format":"int32",
               "description":"Total Men"
            },
			"totalWomen":{
               "type":"integer",
               "format":"int32",
               "description":"Total Women"
            },
			"totalKids":{
               "type":"integer",
               "format":"int32",
               "description":"Total Kids"
            },
			"totalVisitors":{
               "type":"integer",
               "format":"int32",
               "description":"Total Visitors"
            },
            "groupId":{
               "type":"integer",
               "format":"int32",
               "description":"User Group"
            }
        }
      }
    }
}